requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
Flask-RESTful==0.3.9
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
inflect==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
